yet still yet still more cleaning up qe_wrapper [wip]
